backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
torch>=2.1.0
torchvision>=0.16.0
Pillow>=10.0.0
numpy>=1.26.0
reportlab>=4.1.0
nibabel>=5.2.0
scikit-image>=0.22.0
matplotlib>=3.8.0
scipy>=1.12.0
pydantic>=2.5.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
